requirements.txt
python-telegram-bot==21.6
streamlit>=1.36
APScheduler>=3.10.4
python-dotenv>=1.0.1
pydantic>=2.7
loguru>=0.7.2
tzdata>=2024.1
